refactor: #23 등산로 좌표들을 담는 Geometry 리스트 추가
