+++ src/main/java/com/tm/gogo/domain/FavoriteTrail.java
@@ -1,5 +1,6 @@
 package com.tm.gogo.domain;
 
+import com.tm.gogo.domain.hiking_trail.HikingTrail;
 import com.tm.gogo.domain.member.Member;
 import lombok.NoArgsConstructor;
